CIP-1431 fixed bist.sh defined paths

+++ raw/ska-mid-cbf-talondx-bist/bin/bist.sh
@@ -2,17 +2,17 @@
 
 usage() {
     echo "Usage:"
-    echo -e "-s             Start the BIST systemd service and extract tar file"
-    echo -e "-k             Kill the BIST systemd service immediately, aborting the BIST"
-    echo -e "-r             Run the BIST"
-    echo -e "-m <time>      Modify the BIST systemd start delay time by <time>"
-    echo -e "-t             Show the current BIST systemd start delay time"
-    echo -e "-p             Program the BIST bitstream"
-    echo -e "-x             Extract the tar file at predefined location"
-    echo -e "-f             Publish the results (.csv) of the BIST to influxdb"
-    echo -e "-c             Print the results (.log) of the BIST"
-    echo -e "-h             Display the help message"
-    echo -e "-v             Verify the BIST files are installed correctly"
+    echo "-s             Start the BIST systemd service and extract tar file"
+    echo "-k             Kill the BIST systemd service immediately, aborting the BIST"
+    echo "-r             Run the BIST"
+    echo "-m <time>      Modify the BIST systemd start delay time by <time>"
+    echo "-t             Show the current BIST systemd start delay time"
+    echo "-p             Program the BIST bitstream"
+    echo "-x             Extract the tar file at predefined location"
+    echo "-f             Publish the results (.csv) of the BIST to influxdb"
+    echo "-c             Print the results (.txt) of the BIST"
+    echo "-h             Display the help message"
+    echo "-v             Verify the BIST files are installed correctly"
 }
 
 # path of the BIST source files
@@ -97,13 +97,21 @@ extract_archive() {
     # has a limited set of command. tar options such as --get and --wildcards 
     # does not exist.
     echo "Extracting files from tar to $BIST_ARCHIVE_PATH"
-    # remove the output directory if it exists
-    if [ -d "$BIST_ARCHIVE_PATH" ]; then
-        echo "Deleting previous tar file output directory"
-        rm -rf $BIST_ARCHIVE_PATH
+    if [ -f $BIST_ARCHIVE ]; then
+
+        # remove the output directory if it exists
+        if [ -d "$BIST_ARCHIVE_PATH" ]; then
+            echo "Deleting previous tar file output directory"
+            rm -rf $BIST_ARCHIVE_PATH
+        fi
+
+        mkdir $BIST_ARCHIVE_PATH
+        tar -xvzf $BIST_ARCHIVE -C $BIST_ARCHIVE_PATH
+
+    else
+        echo "$BIST_ARCHIVE not found, aborting..."
+        exit 3
     fi
-    mkdir $BIST_ARCHIVE_PATH
-    tar -xvzf $BIST_ARCHIVE -C $BIST_ARCHIVE_PATH
 }
 
 program_bist_bitstream() {
@@ -115,6 +123,7 @@ program_bist_bitstream() {
     # sanity check
     if [ -z "$dtb" ] || [ -z "$bs_core" ]; then
         echo ".dtb or .rbf file(s) missing. Aborting bitstream programming!"
+        exit 3
     else 
         rmdir $BIST_BITSTREAM_PATH/*
         mkdir $BIST_BITSTREAM_PATH/base
@@ -133,7 +142,7 @@ execute_bist() {
         echo "ipmap or json file(s) missing. Aborting BIST execution!"
         exit 1
     else
-        python3 $BIST_SRC_PATH/run_bist_tests.py $BIST_ARCHIVE_PATH/$json $BIST_SRC_PATH/$ipmap
+        python3 $BIST_SRC_PATH/src/run_bist_tests.py $BIST_ARCHIVE_PATH/$json $BIST_SRC_PATH/$ipmap
     fi
 }
 
@@ -149,11 +158,11 @@ run_bist() {
 
 get_bist_results() {
     # print the results of the BIST
-    if [ -f $BIST_SRC_PATH/tdc_base_bist_logfile.log ]; then
-        echo $BIST_SRC_PATH/tdc_base_bist_logfile.log
+    if [ -f $BIST_SRC_PATH/tdc_base_bist_logfile.txt ]; then
+        echo $BIST_SRC_PATH/src/tdc_base_bist_logfile.txt
         return 0
     else
-        echo "$BIST_SRC_PATH/tdc_base_bist_logfile.log not found"
+        echo "$BIST_SRC_PATH/src/tdc_base_bist_logfile.txt not found"
         exit 1
     fi
 }
@@ -161,11 +170,12 @@ get_bist_results() {
 publish_bist_results_to_influxdb() {
     # publish the results to the influxdb. The python script that runs
     # the BIST is responsible for producing a valid .csv file
-    if [ -f $BIST_SRC_PATH/tdc_base_bist_logfile.csv ]; then
-        echo "Publishing $BIST_SRC_PATH/tdc_base_bist_logfile.csv to influxdb"
-        influx write --bucket bist --file $BIST_SRC_PATH/tdc_base_bist_logfile.csv
+    if [ -f $BIST_SRC_PATH/src/tdc_base_bist_logfile.csv ]; then
+        echo "Publishing $BIST_SRC_PATH/src/tdc_base_bist_logfile.csv to influxdb"
+        influx write --bucket bist --file $BIST_SRC_PATH/src/tdc_base_bist_logfile.csv
         return 0
     else 
+        echo "$BIST_SRC_PATH/src/tdc_base_bist_logfile.csv not found."
         exit 1
     fi
 }
@@ -193,7 +203,7 @@ while getopts ":hskrm:vtpxfc" arg; do
         s)
             start_bist_service
             if [ $? -ne 0 ]; then
-                echo "Error starting the BIST service - $status"
+                echo "Error starting the BIST service"
                 exit 5
             fi
             # also extract the bitstream
@@ -201,8 +211,8 @@ while getopts ":hskrm:vtpxfc" arg; do
             ;;
         k)
             stop_bist_service
-            if [ $? -gt 0 ]; then
-                echo "Error killing the BIST service - $status"
+            if [ $? -ne 0 ]; then
+                echo "Error killing the BIST service"
                 exit 4
             fi
             ;;
